Do not blithely delete the categorical directories if the users fails to give a list argument!!!  Complain and bailout if no listname is found.  (Otherwise, when invoked without arguments, the script was deleting the lists, templates, and archives directories *completely*.)
Additionally, do not delete the archives for a list unless the '-a'
option is passed.  I figure its more often worth keeping the archives
around, even if the list is retired.

Additionally, indicate whether or not the subject directory is
actually there before deleting - some feedback, so users see whether
what they expect is happening...