New privacy (ant-spam) feature - bounce_matching_headers option takes a string with "field: matchexp" lines, and, if set, messages with header fields matching matchexp are held for approval. MailList.parse_matching_header_opt separates the header and matchexp using the first ':' (and trims leading whitespace off of the matchexp), and MailList.HasMatchingHeader() does the check.  And of course, MailList.Post() does the check,
MailList.LogMsg(): given the log category, add an entry to the log
file for that category.  Files all situated in setting of (new) config
opt, LOG_DIR.  (Stored the open files on a list, potentially useful
when running as a long running daemon...)

MailList.ApprovedAddMember(): gets a check for existing membership -
possible when more than one subscription for the same address is
okayed from the same pending batch of subscription requests - before
it was creating multiple subscriptions, and things blew up when
attempts were made to revoke the subscriptions...

Bracketed all potentially file-creating 'open()'s
(MailList.CreateFiles(), .Save(), .Lock(), .LogMsg()) in os.umask
suspension/resumption, so that the files are created with the intended
permissions.

Added a repr which indicates the lock state of the file.  (Helpful for
my exploratory programming.)

Reworded some option descriptions.