ObscureEmail(), UnobscureEmail(): new routines for invertably transforming email addresses so they're not apparent to web spiders, therefore not scrapable by spammers.
The UnobscureEmail transformation should work like an identity
function on already-unobscured addresses, so it can be used with
impunity.

The routines should be used as functional interface to address
obscuring, so the scheme can be changed transparently to the rest of
the system, just by changing the routines.  This should make it easy
to up the ante in the thwarting of spammers, if this catches on.