Respect obscured addresses for user identity, as well as unobscured ones (as long as mm_utils.UnobscureEmail() works like an identity function on not-obscured email addrs).
Also, set replacement <mm-presentable-user> to the obscured address if
address obscuring is on.  (Used by options.html.)

NOTE that the maillist administrator's address is never obscured.
Maillist administrators will have to either present the maillist-owner
address as their admin addr, or otherwise protect themselves if they
wish.

Remove redirection of stderr to file - the script would fail if the
file already existed and was unwritable.  (It may make sense to have
a logging mechanism like LogMsg, but working like a file object, for
this purpose...)

Add a module docstring.