Resolving digest header and footer formats against list.__dict__ (just before posting).
Getting digest header and footer from (new) default settings.

Changed subject of footer message from "Reminder" to "Digest Footer".

(NOTE that the mime seems to be kinda funky - my mua, a fairly current
version of Unix pine, does not recognize it as mime, and it tends to
be pretty good about that kind of thing.)