We're just archiving to a flat file now, leaving the real work to an externally running pipermail.  (Andrew and i are working together to get easy hookup, and we should be able to relaese something of pipermail along with the near release of mailman.)
Moved the creation of the archive file hear, when it's openned, and
protected the open so it doesn't bomb the system if it fails - and a
'system' error is logged to the mailman logs in that situation.

The archive (mbox) file is created in either the
PUBLIC_ARCHIVE_FILE_DIR or PRIVATE_ARCHIVE_FILE_DIR dependig on the
setting of DEFAULT_ARCHIVE_PRIVATE (typically 0, 1 for private
archives).  Pipermail will be set to look in the private and public
archive dirs and do the right thing - with easy means to hook up the
mailman password mechanism for the private archives.