This is the fix i mentioned in my last checkin (not mm_digest).
Use a python-lib tempfile, instead of ginning one up distinguised by
pid.  The problem with the old method is that the same process - eg,
senddigest - may be cycling through many lists, which means that all
the lists would use the same tempfile name.  And deliver, having
opened the file for reading, would get it pulled out from under it
when, eg, senddigest got to having mm_deliver prepare digest for the
next list.  tempfile.mktemp() specifically finds a filename that is
not in use.

I'm leaving in commented-out text for some special membership-
migration notices, for use later today or tomorrow, when i move over
the rest of the lists.