Looks like bounce handling does not work like options advertise - in particular, "Don't remove, notify me" still *does* do the removal. Sigh.
The right way to set this up will take some thought, but i think my
first cut fix will be, when this option is selected, to just disable
the user, and send a note to the admin.  That way, the admin won't
just get extra notes accompanying continuing bounce results for the
user...  I may get to this quick fix soon.

The long term (and fairly attractive) fix is to make a bounce-handling
subsystem which collects bounce messages and classifies them according
to a admin-configurable template, taking actions like disabling or
removing the user, notifying the admin, etc, when admin-configured
thresholds are reached.  The admin interface could be handled very
like the admin-requests mechanism, with a page for bounce
administration and a periodic notice of pending bounce admin stuff.