Changed names of 'closed' and 'auto_subscribe' options!
'closed' => 'private_roster'.  'Closed' was being used to control
whether the roster of list members was to be visible to non-list
members, but is used elsewhere (in mailman explanations for
subscription deferrals, as well as in majordomo and other maillist
system parlance) to refer to whether the subscription policy requires
admin approval or not.

'auto_subscribe' => 'open_subscribe'.  This is the option that
controls whether or not the list is closed - 'open' is more clear than
'auto' - i would have switched to 'closed', but saw no need to have
the name refer to the *much* less common case, and didn't want to
change the default sense of the option.

Added listme.com to the roster of forbidden spam posters.