Added logger and moved SendTextToUser and DeliverToUser functionalities here.
Logger() and StampedLogger() routines provide file-like file logging
mechanisms.  (I reimplmented the list .LogMsg() method based on this
routine.)  This is implemented outside the list infrastructure so it
can be used for, eg, stderr substitution.

 SendTextToUser() and DeliverToUser() implemented here likewise, so
they can be used outside list context (eg, when lists are broken...).

Made AddressesMatch not raise error when one of the address arguments
yields a null domain.

Moved comment about .FindMatchingAddresses() to be the function
docstring.