Added handling for subscription-list presentation request.
call_script() - Encapsulation of hack to so i could use it for the
mmroster script as well as for the options script.

Using StampedLogger (to logs/error) for stderr, for debugging.

Changed error notices to have a header "Error" and a body describing
the problem.

Added modest module doc string.