AddressesMatch(): Do the right thing (in particular, no traceback!) when one of the addresses has no domain.
StampedLogger(): New option manual_reprime makes better for use as
sys.stderr substitute - says to put the timestamp only on first write
or writes just succeeding a .reprime() call - so the timestamp divides
separate multi-line entries.  And in any case, write the
timestamp/label prefix on when at the beginning of a new line.