The double-use of the 'subscribe' cgi script for subscription and edit-options had a hidden problem - entering info in the edit-options "info" field with a carriage return (in netscape, at least) resulted in doing the default subscribe submit, which would get rejected due to lack of email and password.
For now i've just elaborated the check for edit-options to look
additionally (besides checking for the right submit button having been
hit) for the presence of a value in the edit-options info field *and*
the absence of any value in the subscribe email field.  This is enough.

The proper thing would be to separate the functions into two separate
scripts (as i did for the subscriber roster, this morning), but that
will take adding another wrapper, etc - maybe when john gets a chance
to clean up the wrapper situation...  (That's not a prod, john -
really!-)

By the way!  With the recent changes, it's now possible for you guys
to see who actually are the members of the mailman-cvs list (a private
list), via the web.  There really are only three of us - john, barry,
and me.  In case you were curious...

Ken