Implement rudimentary admin-configuration long-description help.  I now have to refine it, but it's simple and it works.
It uses the anchor href "target" option to display the help in a named
alternate window.  I presume that in browsers that do not support this
feature they will simply go to another page, where the user can hit
'back' when they're ready.

Also, right-aligned the short descriptions on the left column of the
table, so they snug up to the value column.