HTMLFormatObject(): Ok, here's a stab at making the error-protection a bit more specific and unobtrusive.
Instead of using a bare except to just silently ignore all encompassed
formatting bugs, invoke .Format() only for items that hasattr
identifies as having the method.  For any other items, return the
string representation of the item.  This way, python objects that are
mistakenly expected to behave like containers or whatever will be
presented as the python objects they are - making it easier to track
down the error, without crashing the presentation process as a whole.
Moreover, it may be a non-error - any object with a string
representation can be put in an htmlformat container, and show
through.