Pushed the inline code at the bottom into a main() routine.
GetListobj(): gets the mailman list object, using a stashed copy if
it's a repeat call.  And safely unlocks the list, if it's locked.

Use a self-contained copy of the form.

I've created an executable mailman-cgi wrapper to take care of the
setuid capability, so i've moved the script to this new location, as
well.