.SaveForDigest():
 - Digest messages now keep all headers except 'received',
   'errors-to', and all 'x-*' ones (and any continuations of these).
 - Mime digest messages now properly (i believe) contain mime
   attachments.  (The key was preserving the original content-type
   header.)
 - It may be that some more headers should be trimmed, ie for the sake
   of the non-mime recipients - but actually, it doesn't look
   cluttered, for the few sample messages i tried.
 - The entries in the table-of-contents have the redundant
   subject-prefix string removed.

Digest: Redundant "Vol" in masthead string removed.

I took a look at rfc 934 (thanks, barry), which is the pre-Mime standard
for encapsulating (nesting) messages in other messages, and it looks
like the key is to have an "encapsulation boundary" that starts with a
"-" dash - looks like mime extended from this.  However, the old rule
for nesting an encapsulated message within another encapsulated
message is to insert a "- " space in front of the original
encapsulation boundary.  Unfortunately, this would break mime
encapsulation, so i'm not implementing this part of the old
burstable-digests standard.  The rest of it *may* be ok, though, in
case anyone has the old-style bursting readers...

Anyway, i've done about as much as i was hoping to do with the digests
format - i don't forsee devoting much more attention to features, just
to ironing out bugs.