Prevent bounce loops when admin-action recipient is itself the bouncing address - before sending the notice, check whether the bouncing address is among the list owner addresses.
And otherwise, include (attach) a copy of the actual bounce notice in
the admin-action message.  Otherwise there is no record the
administrator can examine.

Finally, send both failed and successful admin-action notifications to
the list administrator.  Before, failed action (e.g. inability to
locate the specific user) were sent to the mailman site admin - this
was while i was debugging the bounce handling, but it looks a lot
better now.