Added provision for a bounce message where the email address is on a different line than the cue.  Thus we have to use two separate checks:
 - One regex, 'separate_cue_1', to identify this as an failure notice.
   The flag var, 'check_prospects', is set if this regex triggers.
 - Another regex, 'separate_addr_1', adds the addr to prospects if
   triggered.

Prospects are only used when 'check_prospects' has been set, to ensure
that the message being examined is in fact a failure notice.

(The two hosts i'm dealing, parrot and glyph, happen to run run
berkeley sendmail 8.8.5, which happens to issue failure notices like
this.  I would not be surprised if more recent sendmails did, too...)