Several changes, first the non-controversial ones...
1. Use the standard Python invocation #! line to get the interpreter
   from $PATH

2. Do no direct sys.path munging in this file.  Import the `paths'
   module, which is created during the configure process, and which
   performs all necessary path munging (and exports some useful
   variables too).

3. Remove RCS crud

And some other changes:

ADDALIASES_CMD does not contain a hardcoded absolute path.  Instead
the $prefix is taken from paths.prefix, which is set by the configure
script.

getusername(): New function which attempts to figure out the user's
login name.  It uses, in order $USER, $LOGNAME, then the element 0 of
pwd.getpwduid(), then the string `<unknown>'.  This is used when
creating the aliases file cut-and-paste.

Slightly changed the format of the aliases file cut-and-paste.  Also
added a `created' line which includes today's date as DD-MMM-YYYY and
the name of the user that created the new list.

Some inconsequential code formatting changes.

Some rewording of the mailing list creation text, and a string
delimiter change to make the code more Emacs friendly.