1. Use the standard Python invocation #! line to get the interpreter    from $PATH
2. Do no direct sys.path munging in this file.  Import the `paths'
   module, which is created during the configure process, and which
   performs all necessary path munging (and exports some useful
   variables too).

3. Remove RCS crud

Also:

Removed all hardcoded absolute paths.  Instead, calculate paths using
paths.prefix as the base.

Change ''' TQS to """ so as not to confuse font-lock.

!!!! When calling SendTextToUser(), set the new optional argument
raw=1 so that the text body will not be wrapped/filled.  The checkin
that implements wrapping and filling has not yet been made.  Note that
it would be nice if just the templated text could be wrapped/filled
and the err_msg left alone, but that's currently too hard to separate.