Removed RCS crud
(SUBSCRIPTION_AUTH_TEXT, POSTING_AUTH_TEXT): Some small, but important
formatting changes.  The wrapping/filling algorithm says not to fill
any lines that begin with whitespace, so for the two embedded lists
that shouldn't be filled, start the lines with a couple of spaces.

RefuseRequest(): Call SendTextToUser() with raw=1 so that the text of
the original message will not be wrapped/filled.  Some of the text
should probably be wrapped/filled but that's too hard to separate out
right now.