Several changes:
Removed RCS crud

In any text sent to the user, substitute `maillist' with `mailing list'

SUBSCRIBEACKTEXT: All %s substitutions are replaced with $(varname)s
substitutions with the corresponding dictionary interpolation added in
Deliverer.CreateSubscribeAck().  With too many %s substitutions, it
gets too hard to keep track of the right number of %s's and their
proper substitution.

Deliverer.SendTextToUser(): Added optional raw argument, defaulted to
0 (meaning default is to wrap/fill -- I know KLM and I disagree about
this ;-).  This argument is passed straight through to
mm_utils.SendTextToUser().

Deliverer.DeliverToList(): When calculating the command to os.popen(),
include as the first argument the configure calculated path to the
Python interpreter to use.  It is possible that `python' is not on the
path when this gets executed (or that the wrong interpreter will be
found).

Deliverer.CreateSubscribeAck(): Use dictionary interpolation.  Also,
The header message seems redundant and unnecessary to me, so I removed
it.