A useful (IMO) change:
The mail wrapper gid and the cgi wrapper gid are now cached values.
This means the first time configure calculates these gid's, it caches
them in config.cache and re-uses them if --with-*-gid are not
specified.  However if --with-*-gid is specified, it's value is used
and stored in the cache.

The advantage here is that if you are constantly overriding one of the
GID's (e.g. as I am because group `www' isn't correct for my system),
I only need to do that once, and then it is remember for each
subsequent build without having to remember which group I chose!  But
I can always override it again if I want.