Added GPL
Even if USE_CRYPT is set, catch ImportError when trying to import
crypt.crypt.  If that fails, then we'll still use our md5 wrapper, and
we'll set USE_CRYPT to 0.

Don't import all of the crypt module's attributes into this module
since the only one we mask anyway is crypt.crypt.