Check for existing lock before unlocking.
QUESTION: The MailList.Unlock() method currently raises two different
AttributeErrors when the list is not currently locked (depending on
whether or not the lock was previously created).  This certainly seems
like the wrong exception.

We could make .Unlock() check for the lock and return a better
exception, or we could have it just return if the list is not locked.
I'm inclined towards the latter, but it changes the character of the
list locking interface a wee bit.

I think it ought to be changed, one way or the other.  John, what's
your inclination?

Ken