Utils.TrySMTPDelivery() was being invoked with the from_addr and to_addrs in the wrong order.
(Looks like it would make sense to change the order of the recipient
and sender arguments in equeueMessage to be consistent with other
routines:

def enqueueMessage(the_sender, recip, text):
def TrySMTPDelivery(recipient, sender, text, queue_entry):
    def send(self, to, frm, text, headers=None):
)

And i did get an aberrant behavior for the first message i posted
after i applied the fix, specifically, the message showed up around 10
times.  Subsequent messages only showed up once, which is proper.  It
may be that some quirk in the failure of probably around 10 previous
trials led to "logjam" in the queue that got cleared by the first
success, but i'm unable to confirm that.  Something to watch out for -
i think it would be good to actually exercise the queue mechanism to
some substantial degree before release.