Ok, i've filled in the architecture (now "structure") using stuff from barry's arch.txt.  I've also taken my first complete pass through the document, refining the stuff i already had there (and adding, for example, a fragment from the driver script, exhibiting the handy dandy logging setup).
I've addressed some of barry's written comments and suggestions,
including applying some of the traditional capaitalization i tend to
ignore, and changing "maillist" to "mailing list" when referring to
the social entity - "MailList" refers to the programming class and
instance entities.

As for your "conventional protocols" question, barry - my
understanding is that conventions *are* agreements - explicit or not.
I think this is a good term, but lets scrutinize further.