Began to address some issues john raised, but need help with these.
1) Fill in and/or fix the static typing discussion.  I need your views
   and your verbage - i'm in bad shape today, myself, and i didn't
   have a good handle on this discussion in the first place.

2) John protested that i introduced some new ideas in the conclusion,
   suggesting it's not a good idea to do that.  I agree, and totally
   toned down the stuff about Python's readability.  (Gad, i'm so
   drawn to spew about that.)  My question is, is there other stuff
   that needs attention?

Added a section about what mailman *is*, at the top of the
introduction.  I almost overlooked this - i'm assuming it makes
sense (but i'm in bad shape today - guys, whatever sanity checking and
rectification you can provide today would be appreciated, we need to
get this done in a few hours, and i've run out of steam.)

Assembled some references - if one of you has time and knowledge about
formatting them, i'd appreciate help with that.

Incorporated two of the diagrams - user-ui.jpg and admin-ui.jpg.
Where should the structure diagram go, if it's going to go in?

Did a spell check.