TrySMTPDelivery: The queueing activity had to be more discriminating, or else failures, like bad local recipients, would remain on the queue forever, and cause repeat deliveries.  (I believe this is part of the problem behind the repeats we're seeing - but i'm not convinced it's all of it.)
To fix, i made it so only partcicular exceptions - currently
socket.error, for absent SMTP listener - cause items to be left on
queue.  Otherwise it falls through to a blanket except which discards
the item, leaving a note in the error log to that effect.

***  I think where the info should go back to the maillist structure,
to, e.g., disable the recipient, or whatever.  However, this is
happening in a forked process, so we cannot use an exception, and the
routine itself doesn't, and shouldn't know which the list is.