.SetHeader(): New headers would not show in resulting message unless optional "crush_duplicates" was set to 0, because the msg.dict would get the new entry, but the msg.headers would not.  I've changed the code so that the non-crush_duplicates addition to msg.headers would happen whenever the header is new, regardless of crush_duplicates setting.  (Of course, the newness of the header means there are no duplicates to crush...)
Incidentally!  John, there's a line at the top of this routine whose
purpose i don't understand - i think it was there when i got the code,
so i'm wondering if you know about it.  It's the first line of code in
the routine:

	name = "%s%s" % (name[0], name[1:])

The only thing i can think this would serve is to establish that the
length of the name is at least 1 - seems to me it would be more
efficient and clear to make the check explicit, if that's the purpose,
avoid the assignment, etc.  I suspect i'm missing something here - can
you shed any light??