Add new "list-id:" header to all messages sent out by the list, according to, in accordance with RFC draft-chandhok-listid-02.  This includes all admin messages and postings sent to members, _except_, those postings that already have a list-id header keep the original and don't get the new one.
This basically conforms to the spec.  However, the spec stipulates
that individual users should not be able to specify a list-id, yet
postings from other maillists should, and those other maillists
list-id headers should not be overridden.  Since there is no way to
definitely tell the difference, i'm just respecting the original
list-id.  (It may be that the intent is to add to, rather than defer
to, the existing list-id header if the message is from another list.
If that's the case, we should just change the code to unconditionally
add the header.)