Incorporated DSN provision, so that receiving hosts that support DSN can be instructed to send delivery-status-notices only when the delivery fails, not for intermediate warnings.
This is modelled after the patches dan ohnesorg submitted quite a
while back - but they really were pretty fouled up, particularly when
the receiving hosts doesn't support dsn, so it's really rewritten.
(And only around 15 lines total change, mostly different than the
original patch.)

Once we get bounce handling more able to adapt to various formats, we
might want to reenable warnings to help track unreachability status.
Until then they are a nuisance.  And we should have shifted over to
the python-distribution smtplib.py well before then - as of 1.5.2 it
(will/does) have esmtp support of a sort - and we shouldn't be
dabbling in this garbage.

And yes, somebody slap my hand for adding a new feature - it'll hurt
so good if it means eliminating the goddamn warning notices, the
plague of large list administrators...  (I really wanted to get this
working before midnight - when passwords go out, to hundreds of
python.org mailman maillist subscribers!)