main(): A small kludge to ensure that we use an advertised list, if any is available, for actually sending out the passwords reminders. Otherwise, an unadvertised list may be used (and will be, if there are no advertised ones on the system), putting its name in the list-id header, and so forth.
The proper fix will be to not use a list object, at all, to do the
sends - use Utils.SendTextToUser, or something, directly.  This should
be fixed before full release - we just need to identify the proper
headers - who should be the sender, where should failure notices go,
etc, etc.  Hmm - it would be nice to have a special return address
that would do the bounce processing, but look for and handle the bad
address on *all* the lists to which it belongs...