.ArchiveToMbox(): Remove gratuitous self.Save() - particularly problematic in a forked image of the list - both the parent and the child have the lock, and they're both going to use it!
.ArchiveMail(): Do an os._exit() instead of a return in the case where
ARCHIVE_TO_MBOX is 1 (mbox only).  Otherwise we were returning to a
second version of the caller, and havoc ensued...-)

There is still a problem with not having the final os._exit() ensured
as part of a "try:...finally:" clause, which actually causes
redundancy in the operation of the post script.  The problem is:

 - the post script, for example, calls ArchiveMail() within the try
   section of a try...finally
 - ArchiveMail() may encounter an exception before it hits the os._exit()
 - and thus, the stack frame will be popped back to post(), hitting
   the finally which gets executed.

I'd like to just put the ArchiveMail()'s os._exit() in a finally
clause, to make sure it always happens, but for some reason that is
preventing tracebacks from making their way to the error log.  We need
to determine why that's happening, and do the right thing, otherwise
we *are* getting multiple returns from routine when an exception
occurs.