Mods to work with new archiving scheme.
ROOT is gone.

getListName(): must extract list name differently.

GetListobj(): Interface change!  Allow Errors.MMUnknownListError to
propagate up.  Returning None is not helpful because this will
generally cause an AttributeError later on, and that just masks the
real error, which is that the list name was not found.

main(): when file isn't found, include the true_filename in the
message (for debugging - shouldn't normally happen).