Changes to avoid having to chmod files, which may fail.
Import open_ex() from Mailman.Utils and assign it to open in the
module's globals, so this gets picked up before builtin open.  Saves
rewriting lots of occurances of open(), but could be confusing when
reading a method.  Hmmm...

Changes to the default values for DIRMODE and FILEMODE.

HyperArchive.add_article(): use mkdir() from Mailman.Utils