Changes to avoid having to chmod files, which may fail.
Import open_ex() from Mailman.Utils and assign it to open in the
module's globals, so this gets picked up before builtin open.  Saves
rewriting lots of occurances of open(), but could be confusing when
reading a method.  Hmmm...

__openIndices(): If the `database' directory does not exist, create it
specifically with mode = 02770.  We want o-rx so this directory is not
accessible when the archive is public.  But we must have g+wxs or the
competing processes (mail and web) that try to update these files
(when a message is posted, or web approved after being held), can get
to and write the files.