TOC_entry_template: generate the entire "Text" table cell (see below).
html_TOC(): We're even more sophisticated when generating the Text
cell.  We first look for the .txt.gz file and then the .txt file,
plunking in the link and some extra explanatory text depending on what
we find.  If we find nothing, we suppress the entire cell.  Also, we
calculate the size (approximately) and give an indication in the cell
text.

ALSO: it is realized that nothing special has to be handled to promote
.txt to .txt.gz when gzip `suddenly appears'.  The current algorithm
already handles this.  Note there's not much we can do if it suddenly
disappears (since we can't un-gzip it if the module's not available!).

import sys at top of file instead of multiple places where used.  also
convert all sys.stderr.write()'s to self.message() so they'll honor
the VERBOSE flag.