.InitVars(): Include a comment explaining the bounce_info structure, at least, as best as i can glean.
.RegisterBounce(): Ensure that self.Save() is called if the bounce
history stats change within the routine.  This was not always
happening before, so that changes in the stats could be lost
(depending on whether some other method saved the config somewhere
down the line), leading to inconsistent behavior.  I fixed that, and
added a loop at the top that culls stale entries from the db whenever
a bounce is detected, so we aren't monotonically accumulating bounce
histories over the lifetime of the list!

I also added a docstring and several less cryptic comments, to try to
be clearer about what's going on here.