Here's two drastic policy-level changes to the archive mechanism. First of all, i'm setting the default message archive style to preserve horizontal whitespace, use <pre>, instead of putting a <br> at the beginning of every line.
More deeply, i'm inhibiting all completely-uppercase variables from
being included in the pickled settings for the list.  As with the
VERBOSE problem i addressed in my last checkin, changes to one of the
class default settings at any time *after* the archive object state
was resurrected would be reinstated thence forward - overriding the
defaults, and any other settings.  This is not a good interface for
setting options - and i'm not really offering an alternative, other
than making the defaults come back the next time the list is
reinstated, rather than preserving the changes.  We need to resolve
this - i suppose with an interface that distinguishes permanent from
temporary settings.

In general i've changed the defaults so they preserve more of the
structure of the postings (and included very brief comments presenting
what i could glean of the settings from the code - this code here does
not seem to have been written to be understood).