Fixed another place where new accounts are saved without the password. I added a list.Save() after each section (non-digest-members and members) that had any additions.
I should note that the some partial saving *was* being done, just
before the password entry was added - and probably during the addition
of _every_ member.  This would seem to defeat the speedup benefits,
while also introducing a bug of partial saving.  This should be
rectified one way or the other!