main(): Kludge-around for accessing variable details when cookies are disabled.  The problem: because cookies are disabled, the admin has to authenticate for each access.  We can't change this because otherwise option values could leak.
So when the admin fills in the password and clicks `let me in...', the
form gets POSTed, and even those the variable to get details on is
given in the form's ACTION, cgi.py ignores any QUERY_STRING on a
POST.  Maybe cgi.py should be fixed?  In the meantime, we just watch
for the FieldStorage having a `request_login' field and a QUERY_STRING
environment variable.  In that case, we manually parse out the query
string and use its value if available.