ValidateEmail(): Renamed from ValidEmail().
Cleaned up its interface so that it always raises an exception if an
address problem is uncovered (it used to both raise exceptions and
return 0 or 1).  It now has no useful return value.

Added a re to barf on other unacceptable characters in an email
address, currently: [ ] ( ) < > | ;
Should there be others? The test that were there are retained, but the
domain_parts test was cleaned up (there was a redundant test in
there).

Note that now unqualified addresses (e.g. those that don't contain an
`@' are now DISALLOWED).  This goes partially toward Greg Stein's
suggestion.