SetUserDigest(): When turning digests off, stick the entry for digest_members onto one_last_digest attribute.  This attribute is checked in SendDigest().
SendDigest(): Add the members in one_last_digest attribute to the
digest recipients.  This means that folks who turn digests off at
least don't lose messages.  This is a kludge but it was much harder to
send the digest to the user when they undigest, so this is a workable
compromise.