Lots of changes!
(option_descs, option_info): Provide much more detailed descriptions
of the options.  Also changed `norcv' to `notmetoo' although the
former is still acceptable as an alias.

(options): New global to preserve the order that we want to display
the options in response messages.  These options are in the same order
as shown on the membership management admin page.  The user option
page still needs to be changed to support this order.

MailCommandHandler.__init__(): All private variables are fixed to
start with double underscore (e.g. __dispatch).  Added an __errors
ivar to count the number of AddError() calls made.  A different error
message format is now used, which should be much more concise, and
more informative information is provided at the head of error
messages.

AddToResponse(), AddError(): Both methods now take a final argument
`trunc' which says what column to truncate the text to.  With trunc=0
no truncation (this is useful for when we use Utils.wrap() to wrap the
text).  AddError() also now takes a prefix argument, increments
self.__error, and essentially calls AddToResponse().

ParseMailCommands(): Some stylistic changes.  The line being processed
is not placed in the response buffer unless we know it is a valid
command.  If it's an invalid command, we print the error message more
concisely.

When `end' is found, include in the response all the following lines
that are ignored.

If the number of errors exceeds 5, then stop processing the message
immediately (don't keep processing more lines).

If an unexpected error occurs with a valid command (one that is
dispatched), we catch this and include the traceback in the response
message.  We also send the traceback to the list administrator -- I
wonder if this should be the mailman-owner???  We also log the
exception.

In the error preamble, include human contact information.

Several of the Process*Cmd()'s have been modified to work with this
new framework, and to provide much more meaningful response messages.